improve code, fix some bugs and start/stop machine with one command

+++ provision.sh
@@ -1,22 +1,37 @@
 #!/bin/bash
 
 # create new user
-sudo useradd -m -s /bin/bash -U $1 -c $1 -p $(openssl passwd -crypt $2)
+if [ ! `sed -n "/^$1/p" /etc/passwd` ]; then
+  sudo useradd -m -s /bin/bash -U $1 -c $1 -p $(openssl passwd -crypt $2)
 
-# copy authorized key
-sudo cp -pr /home/vagrant/.ssh /home/$1
-sudo cp -pr /home/vagrant/.ssh /root
+  # .ssh dir
+  sudo mkdir -p /home/$1/.ssh
+  sudo chmod 700 /home/$1/.ssh
 
-# change onwership
-sudo chown -R $1:$1 /home/$1
-sudo chown -R root:root /root
+  # .ssh auth key
+  sudo echo "$4" >> /home/$1/.ssh/authorized_keys
+  sudo chmod 600 /home/$1/.ssh/authorized_keys
 
-# change root password
-sudo echo "root:$3" | sudo chpasswd
+  # config
+  # echo 'Host 192.168.55.66' >> /home/$1/.ssh/config
+  # echo 'StrictHostKeyChecking no' >> /home/$1/.ssh/config
+  # echo 'UserKnownHostsFile /dev/null' >> /home/$1/.ssh/config
+  # sudo chmod 600 /home/$1/.ssh/config
 
-# authorized without passwd
-sudo echo "%$1 ALL=(ALL) NOPASSWD: ALL" > /etc/sudoers.d/$1
+  # copy authorized key
+  sudo cp -pr /home/$1/.ssh /root
 
-# delete existed users
-sudo userdel -rf vagrant >/dev/null 2>&1
-sudo userdel -rf ubuntu >/dev/null 2>&1
+  # change onwership
+  sudo chown -R $1:$1 /home/$1
+  sudo chown -R root:root /root
+
+  # change root password
+  sudo echo "root:$3" | sudo chpasswd
+
+  # authorized without passwd
+  sudo echo "%$1 ALL=(ALL) NOPASSWD: ALL" > /etc/sudoers.d/$1
+
+  # delete existed users
+  sudo userdel -rf vagrant >/dev/null 2>&1
+  sudo userdel -rf ubuntu >/dev/null 2>&1
+fi
